+++ src/plugins/demo_nodeditor_nodes/Sources/FilePlayback/FilePlaybackModel.h
@@ -3,6 +3,7 @@
 
 #include "NodeDataTypes/SampledData.h"
 #include "FilePlaybackWidget.h"
+#include "shared/IStoppable.h"
 
 #include <QtNodes/NodeDelegateModel>
 
@@ -24,7 +25,7 @@
  * timer runs only while the user pressed Play AND at least one output
  * connection exists; removing the last connection auto-stops the playback.
  */
-class FilePlaybackModel : public QtNodes::NodeDelegateModel
+class FilePlaybackModel : public QtNodes::NodeDelegateModel, public Daqster::IStoppable
 {
     Q_OBJECT