+++ src/plugins/demo_nodeditor_nodes/Sources/GpuMonitor/GpuMonitorModel.h
@@ -4,6 +4,7 @@
 #include "GpuMonitorEngine.h"
 #include "GpuMonitorWidget.h"
 #include "NodeDataTypes/SampledData.h"
+#include "shared/IStoppable.h"
 
 #include <QtNodes/NodeDelegateModel>
 
@@ -18,7 +19,7 @@
  * SampledStreamDescriptor (domain="gpu", 6 FLOAT32 channels) and emits
  * dataUpdated(0). Polling is gated on output connection count (auto start/stop).
  */
-class GpuMonitorModel : public QtNodes::NodeDelegateModel
+class GpuMonitorModel : public QtNodes::NodeDelegateModel, public Daqster::IStoppable
 {
     Q_OBJECT