+++ src/plugins/demo_nodeditor_nodes/Sources/Pcap/PcapModel.h
@@ -4,6 +4,7 @@
 #include "PcapEngine.h"
 #include "PcapWidget.h"
 #include "NodeDataTypes/SampledData.h"
+#include "shared/IStoppable.h"
 
 #include <QtNodes/NodeDelegateModel>
 
@@ -21,7 +22,7 @@
  * and emits dataUpdated(0). Capture is gated on output connection count
  * (auto start/stop) and user Start/Stop.
  */
-class PcapModel : public QtNodes::NodeDelegateModel
+class PcapModel : public QtNodes::NodeDelegateModel, public Daqster::IStoppable
 {
     Q_OBJECT